% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_dependencies.R
\name{report_dependencies}
\alias{report_dependencies}
\title{generate a csv file of the phmxversecore package, including installation details}
\usage{
report_dependencies()
}
\description{
generate a csv file of the phmxversecore package, including installation details
}

